+++ drivers/postgres/postgres.go
@@ -9,6 +9,10 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/neilotoole/sq/libsq/core/retry"
+	"github.com/xo/dburl"
+
 	pgx "github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/stdlib"
 
@@ -773,4 +777,68 @@ func (d *driveri) RecordMeta(ctx context.Context, colTypes []*sql.ColumnType) (
 	}
 
 	return recMeta, mungeFn, nil
+}
+
+// getPoolConfig returns the native postgres [*pgxpool.Config] for src, applying
+// src's fields, such as [source.Source.Catalog] as appropriate. If
+// includeConnTimeout is true, then 'connect_timeout' is included in the
+// returned config; this is provided as an option, because the connection
+// timeout is sometimes better handled via [context.WithTimeout].
+func getPoolConfig(src *source.Source, includeConnTimeout bool) (*pgxpool.Config, error) {
+	poolCfg, err := pgxpool.ParseConfig(src.Location)
+	if err != nil {
+		return nil, errw(err)
+	}
+
+	if src.Catalog != "" && src.Catalog != poolCfg.ConnConfig.Database {
+		// The catalog differs from the database in the connection string.
+		// OOTB, Postgres doesn't support cross-database references. So,
+		// we'll need to change the connection string to use the catalog
+		// as the database. Note that we don't modify src.Location, but it's
+		// not entirely clear if that's the correct approach. Are there any
+		// downsides to modifying it (as long as the modified Location is not
+		// persisted back to config)?
+		var u *dburl.URL
+		if u, err = dburl.Parse(src.Location); err != nil {
+			return nil, errw(err)
+		}
+
+		u.Path = src.Catalog
+		connStr := u.String()
+		poolCfg, err = pgxpool.ParseConfig(connStr)
+		if err != nil {
+			return nil, errw(err)
+		}
+	}
+
+	if includeConnTimeout {
+		srcTimeout := driver.OptConnOpenTimeout.Get(src.Options)
+		// Only set connect_timeout if it's non-zero and differs from the
+		// already-configured value.
+		// REVISIT: We should actually always set it, otherwise the user's
+		// envar PGCONNECT_TIMEOUT may override it?
+
+		if srcTimeout > 0 || poolCfg.ConnConfig.ConnectTimeout != srcTimeout {
+			var u *dburl.URL
+			if u, err = dburl.Parse(poolCfg.ConnString()); err != nil {
+				return nil, errw(err)
+			}
+
+			q := u.Query()
+			q.Set("connect_timeout", strconv.Itoa(int(srcTimeout.Seconds())))
+			u.RawQuery = q.Encode()
+			poolCfg, err = pgxpool.ParseConfig(u.String())
+			if err != nil {
+				return nil, errw(err)
+			}
+		}
+	}
+
+	return poolCfg, nil
+}
+
+// doRetry executes fn with retry on isErrTooManyConnections.
+func doRetry(ctx context.Context, fn func() error) error {
+	maxRetryInterval := driver.OptMaxRetryInterval.Get(options.FromContext(ctx))
+	return retry.Do(ctx, maxRetryInterval, fn, isErrTooManyConnections)
 }
